requirements.txt
apache-airflow==2.3.2
SQLAlchemy==1.4.9
apache-airflow-providers-postgres
apache-airflow-providers-common-sql
psycopg2-binary>=2.9.9